requirements.txt
# Core dependencies
openai~=1.15
anthropic~=0.21
chromadb~=0.4
langgraph~=0.0.26
langchain~=0.1.0
fastapi~=0.109
uvicorn[standard]~=0.27
redis[hiredis]~=5.0
python-dotenv~=1.0
pydantic~=2.5
pydantic-settings~=2.1
watchdog~=3.0

# Utilities
tiktoken~=0.5
httpx~=0.26
tenacity~=8.2
structlog~=24.1
rich~=13.7
gitpython~=3.1
numpy~=1.26

# Development
pytest~=7.4
pytest-asyncio~=0.21
pytest-cov~=4.1
black~=23.12
ruff~=0.1.9
mypy~=1.8
pre-commit~=3.6